Added existence checks and rearranged input handling
Incorporated existence checks for the Nmap file and also for the template and services file, offering to download them if they do not exist and exiting if the user refuses to. Reorganised the input handling, most significantly placing it at the bottom of the file. Added update method for the template file.